Add word, post history and exam history modules; update post, topic, and subtopic features; refactor DTOs and DAOs

+++ src/main/java/com/pbl3/service/SubTopicService.java
@@ -24,6 +24,9 @@ public class SubTopicService implements ServiceInterface<SubTopic>{
 
     @Override
     public int delete(int id) {
+        if(id == 0) {
+            return -1;
+        }
         return subTopicDAO.delete(id);
     }
