+++ src/main/java/com/pbl3/service/TopicService.java
@@ -24,6 +24,9 @@ public class TopicService implements ServiceInterface<Topic>{
 
     @Override
     public int delete(int id) {
+        if(id == 0) {
+            return -1;
+        }
         return topicDAO.delete(id);
     }
